Fix calculation of before change style for CSS transitions
This patch performs a lazy calculation of the before change style once
it has been determined that a transition is being retargeted. In doing
so, the current position is correctly updated to reflect changes made
via the web-animation API.  The path also addresses a bug in the
calculation of current time, in the case of a paused or play-pending
animation.

Bug: 1082401, 888661, 547609

Change-Id: I4b76879d840b482da8ebf23c1aad41b881fafce8